requirements/dev.txt
appdirs==1.4.4; python_full_version >= "3.6.2"
appnope==0.1.2; sys_platform == "darwin" and python_version >= "3.7"
asgiref==3.3.4; python_version >= "3.6"
atomicwrites==1.4.0; python_version >= "3.6" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.6" and python_full_version >= "3.4.0"
attrs==21.2.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
backcall==0.2.0; python_version >= "3.7"
black==21.6b0; python_full_version >= "3.6.2"
certifi==2021.5.30; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
cfgv==3.3.0; python_full_version >= "3.6.1"
chardet==4.0.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
click==8.0.1; python_version >= "3.6" and python_full_version >= "3.6.2"
colorama==0.4.4; sys_platform == "win32" and python_version >= "3.7" and python_full_version >= "3.6.2" and platform_system == "Windows" and (python_version >= "3.6" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.6" and python_full_version >= "3.5.0")
coreapi==2.3.3; python_version >= "3.6"
coreschema==0.0.4; python_version >= "3.6"
coverage==5.5; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4"
decorator==5.0.9; python_version >= "3.7"
distlib==0.3.2; python_full_version >= "3.6.1"
django-colorfield==0.4.1
django-cors-headers==3.7.0; python_version >= "3.6"
django-coverage-plugin==2.0.0
django-debug-toolbar==3.2.1; python_version >= "3.6"
django-extensions==3.1.3; python_version >= "3.6"
django-filter==2.4.0; python_version >= "3.5"
django==3.2.4; python_version >= "3.6"
djangorestframework-simplejwt==4.7.1; python_version >= "3.7"
djangorestframework==3.12.4; python_version >= "3.5"
drf-yasg==1.20.0; python_version >= "3.6"
factory-boy==3.2.0; python_version >= "3.6"
faker==8.8.0; python_version >= "3.6"
filelock==3.0.12; python_full_version >= "3.6.1"
flake8==3.9.2; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
gunicorn==20.1.0; python_version >= "3.5"
identify==2.2.10; python_full_version >= "3.6.1"
idna==2.10; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
inflection==0.5.1; python_version >= "3.6"
iniconfig==1.1.1; python_version >= "3.6"
ipython-genutils==0.2.0; python_version >= "3.7"
ipython==7.24.1; python_version >= "3.7"
isort==5.8.0; python_version >= "3.6" and python_version < "4.0"
itypes==1.2.0; python_version >= "3.6"
jedi==0.18.0; python_version >= "3.7"
jinja2==3.0.1; python_version >= "3.6"
markupsafe==2.0.1; python_version >= "3.6"
matplotlib-inline==0.1.2; python_version >= "3.7"
mccabe==0.6.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
mypy-extensions==0.4.3; python_full_version >= "3.6.2"
nodeenv==1.6.0; python_full_version >= "3.6.1"
packaging==20.9; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
parso==0.8.2; python_version >= "3.7"
pathspec==0.8.1; python_full_version >= "3.6.2"
pexpect==4.8.0; sys_platform != "win32" and python_version >= "3.7"
pickleshare==0.7.5; python_version >= "3.7"
pillow==8.2.0; python_version >= "3.6"
pluggy==0.13.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
pre-commit==2.13.0; python_full_version >= "3.6.1"
prompt-toolkit==3.0.19; python_full_version >= "3.6.1" and python_version >= "3.7"
psycopg2-binary==2.9; python_version >= "3.6"
ptyprocess==0.7.0; sys_platform != "win32" and python_version >= "3.7"
py==1.10.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
pycodestyle==2.7.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
pyflakes==2.3.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
pygments==2.9.0; python_version >= "3.7"
pyjwt==2.1.0; python_version >= "3.7"
pyparsing==2.4.7; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
pytest-django==4.4.0; python_version >= "3.5"
pytest==6.2.4; python_version >= "3.6"
python-dateutil==2.8.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
python-environ==0.4.54
pytz==2021.1; python_version >= "3.6"
pyyaml==5.4.1; python_full_version >= "3.6.1"
regex==2021.4.4; python_full_version >= "3.6.2"
requests==2.25.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
ruamel.yaml.clib==0.2.2; platform_python_implementation == "CPython" and python_version < "3.10" and python_version >= "3.6"
ruamel.yaml==0.17.9; python_version >= "3.6"
six==1.16.0; python_full_version >= "3.6.1" and python_version >= "3.6"
sqlparse==0.4.1; python_version >= "3.6"
text-unidecode==1.3; python_version >= "3.6"
toml==0.10.2; python_full_version >= "3.6.2" and python_version >= "3.6"
traitlets==5.0.5; python_version >= "3.7"
uritemplate==3.0.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
urllib3==1.26.5; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4" and python_version >= "3.6"
virtualenv==20.4.7; python_full_version >= "3.6.1"
wcwidth==0.2.5; python_full_version >= "3.6.1" and python_version >= "3.7"
